分出了三种角色的控制器
