+++ .idea/compiler.xml
@@ -7,10 +7,12 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="dbtest" />
+        <module name="ExaminationSystem" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
       <module name="dbtest" target="1.6" />
+      <module name="ExaminationSystem" target="1.6" />
     </bytecodeTargetLevel>
   </component>
 </project>
